app/src/main/java/nl/cameldevstudio/innavo/models/InnavoLocation.java
package nl.cameldevstudio.innavo.models;

import java.util.Objects;

/**
 * The class represents coordinates of a building.
 */
public class InnavoLocation {
    private double longitude;
    private double latitude;

    public InnavoLocation(double latitude, double longitude) {
        this.longitude = longitude;
        this.latitude = latitude;
    }

    public double getLongitude() {
        return longitude;
    }

    public double getLatitude() {
        return latitude;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        InnavoLocation that = (InnavoLocation) o;
        return Double.compare(that.longitude, longitude) == 0 &&
                Double.compare(that.latitude, latitude) == 0;
    }

    @Override
    public int hashCode() {
        return Objects.hash(longitude, latitude);
    }
}
